extras/watch_vpn_dns/requirements.txt
ConfigArgParse==1.2.3
dnspython==2.0.0
